Refresh dropdown on click
