+++ src/renderer/components/ProcedureDropdown.tsx
@@ -1,19 +1,27 @@
-import { useState } from 'react';
 import { Dropdown } from 'react-bootstrap';
+import { useEffect, useState } from 'react';
 
-interface IProcedureDropdownProps {
-  procedures: string[];
-}
-
-const ProcedureDropdown = ({ procedures }: IProcedureDropdownProps) => {
+const ProcedureDropdown = () => {
   const defaultVal = 'View Stored Procedures';
   const [value, setValue] = useState<string>(defaultVal);
+
+  const proceduresDefault = ['test1', 'test2', 'alexiscool'];
+  const [procedures, setProcedures] = useState<string[]>();
+
+  // useEffect(() => {}, []);
+
+  const fetch = async () => {
+    const newProcs = await window.procedures.ipcRenderer.fetchProcedures();
+    setProcedures(newProcs.get('React'));
+  };
+  fetch();
+
   return (
     <Dropdown className="dropdown">
       <Dropdown.Toggle variant="primary">{value}</Dropdown.Toggle>
 
-      <Dropdown.Menu>
-        {procedures.map((procedure: string) => {
+      <Dropdown.Menu onClick={() => fetch()}>
+        {(procedures || proceduresDefault).map((procedure: string) => {
           return (
             <Dropdown.Item onClick={() => setValue(procedure)}>
               {procedure}
